requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.0
redis==5.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
langcodes==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
cachetools==5.3.2
